+++ Api/CatalogRule.php
@@ -241,6 +241,8 @@ class CatalogRule implements JsonSerializable
     }
 
     /**
+     * Normalize object data for JSON serialization.
+     *
      * @return array
      */
     public function jsonSerialize(): array